Review and potential further dev on content
@abibat-adesina I have updated the review, great work!

I did add in the form of comment (hidden) line 47-53 of this new PR update a couple of comments as user perspective, perhaps these can be developed more or provide links to where to find that specific information.

@wrnu thoughts?

Anyway, if you wish to add further information, once that is done send it back to me again to review it as well. Please note I am away from Dec 22, 2023 - Jan 2, 2024